CRC_CALC.c edited online with Bitbucket

+++ Src/CRC_CALC.c
@@ -6,7 +6,7 @@ const uint32_t BUFFER_SIZE = (uint32_t)(6);
 uint8_t calculate_base_CRC( uint8_t * data_bytes )
 {
     uint8_t crc_value = (uint8_t)HAL_CRC_Calculate(&crc_handle, (uint32_t*)(data_bytes), BUFFER_SIZE);
-    return (crc_value ^ 0xff);
+    return (uint8_t)(crc_value ^ 0xff);
 }
 
 void init_crc_handle()
